test(LineTest): SectionsTest를 LineTest로 옮김
